requirements-ml.txt
torch
torchvision
ultralytics
transformers
opencv-python-headless
accelerate
pillow
numpy<2
